fix: set the ANCHORE_GITHUB_TOKEN to "default-unset" instead of null, and add tests
Having a non-null string value ensures proper rendering of the value
into the secret and is a no-op when the
anchoreConfig.feeds.drivers.github.enabled property is false.

When a user enables that flag, it is expected that the token value is
set to a non-default value in order for the driver to work.

Signed-off-by: Zach Hill <zach@anchore.com>